pyproject.toml
[project]
name = "ch10-1553-flightgen"
version = "1.0.0"
description = "Generate realistic CH10 files with 1553 flight test data"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pychapter10>=1.1.19",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pyinstaller>=5.13.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
ch10gen = "ch10gen.__main__:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["ch10gen*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "optional: marks tests that require optional dependencies",
    "compat: marks tests that require PyChapter10 compatibility mode",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]